Split monolithic InfoPanel module into sub-modules
